feat: make header tak whole screen and sticky
